audio channels 6 -> 2
Unless you have a digital receiver that decodes AAC MultiChannel
surround, having 6 channels makes no sense. In fact, it makes it worse
because it gets truncated to stereo by the receiving player (Apple TV,
iPhone) and has to deal with truncating the channels itself, which it
does a terrible job.

So, let's let FFMPEG just do that conversion for us.